test: add test for NkHomeApp
