+++ lib/nk_home_app.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/widgets.dart';
+import 'package:go_router/go_router.dart';
+
+class NkHomeApp extends StatelessWidget {
+  static const title = 'NK Home';
+
+  const NkHomeApp({
+    Key? key,
+    required this.router,
+  }) : super(key: key);
+
+  final GoRouter router;
+
+  @override
+  Widget build(BuildContext context) {
+    return WidgetsApp.router(
+      routeInformationParser: router.routeInformationParser,
+      routerDelegate: router.routerDelegate,
+      title: title,
+      color: const Color(0xFFFFFFFF),
+    );
+  }
+}
